Create Readme.md file
The aim of this project focuses on connecting the dots between the patients & the hospitals to dissolve the tough boundary between them. We aim to resolve this non-linearity in the hierarchy by creating a mobile application based system, which will be centralized. As the private healthcare sector is highly unregulated, this project will bring uniformity in healthcare sectors.